- Improved 64bit left shift and removed BootLoader project
