+++ Source/MOSA.sln
@@ -136,8 +136,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mosa.Workspace.UnitTest.Deb
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mosa.UnitTests.x86", "Mosa.UnitTests.x86\Mosa.UnitTests.x86.csproj", "{44725E46-513F-4CFB-8196-2191399C514B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mosa.BootLoader.x86", "Mosa.BootLoader.x86\Mosa.BootLoader.x86.csproj", "{EA460394-B868-45FB-8B7F-4096550B3113}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mosa.Tool.Disassembler.Intel", "Mosa.Tool.Disassembler.Intel\Mosa.Tool.Disassembler.Intel.csproj", "{D9A2A321-F63A-495D-905F-EF7C84E4BB81}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mosa.Utility.RSP", "Mosa.Utility.RSP\Mosa.Utility.RSP.csproj", "{5B810793-78E2-413C-CC05-FCDD12B7DB04}"
@@ -188,7 +186,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mosa.Tools.Package", "Mosa.
 		{F445DB8E-E25A-41EF-A158-C742F14E94E4} = {F445DB8E-E25A-41EF-A158-C742F14E94E4}
 		{5B810793-78E2-413C-AA05-FCDD12B7DB04} = {5B810793-78E2-413C-AA05-FCDD12B7DB04}
 		{5B810793-78E2-413C-CC05-FCDD12B7DB04} = {5B810793-78E2-413C-CC05-FCDD12B7DB04}
-		{EA460394-B868-45FB-8B7F-4096550B3113} = {EA460394-B868-45FB-8B7F-4096550B3113}
 		{577B8294-334B-4E02-BE1A-CBA8ABAF8939} = {577B8294-334B-4E02-BE1A-CBA8ABAF8939}
 		{577B8294-334B-4E02-BE3A-CBA8ABAF8939} = {577B8294-334B-4E02-BE3A-CBA8ABAF8939}
 		{A47B7998-4EE8-4AF8-9373-C664FDEDDC65} = {A47B7998-4EE8-4AF8-9373-C664FDEDDC65}
@@ -707,18 +704,6 @@ Global
 		{44725E46-513F-4CFB-8196-2191399C514B}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{44725E46-513F-4CFB-8196-2191399C514B}.Release|x86.ActiveCfg = Release|Any CPU
 		{44725E46-513F-4CFB-8196-2191399C514B}.Release|x86.Build.0 = Release|Any CPU
-		{EA460394-B868-45FB-8B7F-4096550B3113}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EA460394-B868-45FB-8B7F-4096550B3113}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EA460394-B868-45FB-8B7F-4096550B3113}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{EA460394-B868-45FB-8B7F-4096550B3113}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{EA460394-B868-45FB-8B7F-4096550B3113}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{EA460394-B868-45FB-8B7F-4096550B3113}.Debug|x86.Build.0 = Debug|Any CPU
-		{EA460394-B868-45FB-8B7F-4096550B3113}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EA460394-B868-45FB-8B7F-4096550B3113}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EA460394-B868-45FB-8B7F-4096550B3113}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{EA460394-B868-45FB-8B7F-4096550B3113}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{EA460394-B868-45FB-8B7F-4096550B3113}.Release|x86.ActiveCfg = Release|Any CPU
-		{EA460394-B868-45FB-8B7F-4096550B3113}.Release|x86.Build.0 = Release|Any CPU
 		{D9A2A321-F63A-495D-905F-EF7C84E4BB81}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D9A2A321-F63A-495D-905F-EF7C84E4BB81}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D9A2A321-F63A-495D-905F-EF7C84E4BB81}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
@@ -888,7 +873,6 @@ Global
 		{2A98B95B-B900-485C-85F7-7CD2944612C0} = {88D3D1D0-85D2-4301-A9EA-3F964CB05415}
 		{1717C7EB-06ED-434B-BA3C-E46C9644EEDF} = {AC7C3CDA-E0E8-408C-BB32-DBCBB905E3E1}
 		{44725E46-513F-4CFB-8196-2191399C514B} = {1AB7FA6A-829C-4DCF-BCA6-A27156D8DA26}
-		{EA460394-B868-45FB-8B7F-4096550B3113} = {1AB7FA6A-829C-4DCF-BCA6-A27156D8DA26}
 		{D9A2A321-F63A-495D-905F-EF7C84E4BB81} = {D032B24A-CE3A-4881-BACE-CC4FE0AFD69D}
 		{5B810793-78E2-413C-CC05-FCDD12B7DB04} = {90065B0F-1BFE-40D8-AED5-11096B2535B0}
 		{1717C7EB-06ED-434B-BA3C-E4FF9644EEDF} = {AC7C3CDA-E0E8-408C-BB32-DBCBB905E3E1}
