--- Source/Mosa.BootLoader.x86/ReadMe.txt
@@ -1,1 +1,0 @@
-This is an incomplete, experimental project.
