--- Tests/Build-BootLoader.bat
@@ -1,1 +1,0 @@
-..\bin\Mosa.Tool.Launcher.exe --q --a --qemu -map --asm --launch-off --all-optimization-off --base 0x00200000 --dest BootImage ..\bin\Mosa.BootLoader.x86.exe
